refactor(public-www): single currency prefix helper in format.ts
Export normalizeCurrencyPrefixForDisplay for calendar/event API strings;
formatCurrencyDisplayPrefix delegates HKD to it. events-data uses the
shared helper instead of a local duplicate.
