+++ apps/public_www/tests/lib/format.test.ts
@@ -6,6 +6,7 @@ import {
   formatCurrencyDisplayPrefix,
   formatCurrencyHkd,
   formatPartDateTimeLabel,
+  normalizeCurrencyPrefixForDisplay,
   parseCohortValue,
 } from '@/lib/format';
 
@@ -24,6 +25,15 @@ describe('formatCurrencyHkd', () => {
   it('formats HKD values using locale-aware output', () => {
     expect(formatCurrencyHkd(9000, 'zh-CN')).toBe('HK$9,000');
     expect(formatCurrencyHkd(9000, 'zh-HK')).toBe('HK$9,000');
+  });
+});
+
+describe('normalizeCurrencyPrefixForDisplay', () => {
+  it('maps HKD to HK$ and passes through other prefixes', () => {
+    expect(normalizeCurrencyPrefixForDisplay('HKD')).toBe('HK$');
+    expect(normalizeCurrencyPrefixForDisplay(' hkd ')).toBe('HK$');
+    expect(normalizeCurrencyPrefixForDisplay('HK$')).toBe('HK$');
+    expect(normalizeCurrencyPrefixForDisplay('  ')).toBeUndefined();
   });
 });
